fix: undefined behavior in BMP ReadLong (signed shift overflow)
Found by fuzzing the decoders under UBSan: ReadLong assembled a 32-bit value
with (c & 0xff) << 24, which overflows signed int when bit 7 of the byte is
set (UB). Perform the shifts in unsigned and convert back to int.

+++ src/IBMP.c
@@ -53,13 +53,14 @@ static int ReadLong (FILE *fp, int *ret)
   *ret = (c&0xff);
   c = fgetc(fp);
   if (c == EOF) return 0;
-  *ret |= (c&0xff) << 8;
+  *ret |= (int)((unsigned int)(c&0xff) << 8);
   c = fgetc(fp);
   if (c == EOF) return 0;
-  *ret |= (c&0xff) << 16;
+  *ret |= (int)((unsigned int)(c&0xff) << 16);
   c = fgetc(fp);
   if (c == EOF) return 0;
-  *ret |= (c&0xff) << 24;
+  /* shift as unsigned: (c&0xff)<<24 would overflow signed int when bit 7 set */
+  *ret |= (int)((unsigned int)(c&0xff) << 24);
   return 1;
 }
